Move some objects from NDLC.CLI to NDLC
