+++ NDLC.CLI/DLC/StartDLCCommand.cs
@@ -12,7 +12,8 @@ using System.CommandLine.Parsing;
 using System.ComponentModel.Design;
 using System.Text;
 using System.Threading.Tasks;
-using static NDLC.CLI.Repository.DLCState;
+using NDLC.Infrastructure;
+using static NDLC.Infrastructure.Repository.DLCState;
 
 namespace NDLC.CLI.DLC
 {
