+++ NDLC/Infrastructure/JsonHelpers.cs
@@ -1,0 +1,13 @@
+using NDLC.Messages;
+using Newtonsoft.Json.Linq;
+
+namespace NDLC.Infrastructure
+{
+    public static class JsonHelpers
+    {
+		public static JObject ExportStateJObject(this DLCTransactionBuilder builder)
+		{
+			return JObject.Parse(builder.ExportState());
+		}
+    }
+}
