+++ NDLC/Infrastructure/NameRepository.cs
@@ -1,15 +1,12 @@
-﻿using NBitcoin;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Serialization;
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.Collections.Specialized;
 using System.IO;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Serialization;
 
-namespace NDLC.CLI
+namespace NDLC.Infrastructure
 {
 	public class NameRepository
 	{
